+++ includes/blocks/table-of-contents/src/edit.js
@@ -1,11 +1,18 @@
 import {__} from '@wordpress/i18n';
 import {useBlockProps} from '@wordpress/block-editor';
+import {SelectControl} from '@wordpress/components';
 import {getLanguageControls, getDisplayValueFromArray} from '../../blocks';
 import './style.css';
 
+const ALLOWED_TAGS = ['div', 'p', 'span', 'button', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6'];
+
 export default function Edit({attributes, setAttributes, isSelected}) {
 
-    const {labels} = attributes;
+    let {className, labels, htmlTag} = attributes;
+    if(className === undefined) {
+        className = 'is-style-oes-default';
+    }
+    if (!ALLOWED_TAGS.includes(htmlTag)) htmlTag = 'div';
 
     if (isSelected) {
         return (
@@ -14,13 +21,39 @@ export default function Edit({attributes, setAttributes, isSelected}) {
                     <div className="components-placeholder__label">{__('Table of Contents', 'oes')}</div>
                     <div className="oes-block-subheader">{__('Header', 'oes')}</div>
                     {getLanguageControls(labels, setAttributes)}
+                    <SelectControl
+                        label={__('HTML Tag', 'oes')}
+                        options={[
+                            {label: 'Block', value: 'div'},
+                            {label: 'Paragraph', value: 'p'},
+                            {label: 'Span', value: 'span'},
+                            {label: 'Button', value: 'button'},
+                            {label: 'H1', value: 'h1'},
+                            {label: 'H2', value: 'h2'},
+                            {label: 'H3', value: 'h3'},
+                            {label: 'H4', value: 'h4'},
+                            {label: 'H5', value: 'h5'},
+                            {label: 'H6', value: 'h6'}
+                        ]}
+                        value={htmlTag}
+                        help={__('The html tag defines the presentation of the content.', 'oes')}
+                        onChange={(val) => {
+                            setAttributes({
+                                htmlTag: String(val)
+                            })
+                        }}
+                    />
                 </div>
             </div>
         );
     } else {
+
+
+        const Tag = htmlTag;
+
         return (
-            <div {...useBlockProps()}>
-                <h2 className="oes-content-table-header">{getDisplayValueFromArray(labels, '')}</h2>
+            <div {...useBlockProps()} className={className}>
+                <Tag className="oes-content-table-header">{getDisplayValueFromArray(labels, '')}</Tag>
                 <ul className="oes-table-of-contents oes-vertical-list">
                     <li className="oes-toc-header2 oes-toc-anchor"><a>{__('Lorem Ipsum', 'oes')}</a></li>
                     <li className="oes-toc-header2 oes-toc-anchor"><a>{__('Dolor sit Amet', 'oes')}</a></li>